Move classes to right folders
