+++ RateLimiter/Models/Request.cs
@@ -1,10 +1,10 @@
 
 using System;
 
-namespace RateLimiter;
+namespace RateLimiter.Models;
 public class Request
 {
-    public string Resource { get; init;} = null!;
+    public string Resource { get; init; } = null!;
     public Token Token { get; set; } = null!;
 
     public DateTime Timestamp { get; set; }
