+++ RateLimiter/Models/Token.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace RateLimiter.Models;
+public class Token
+{
+    public Guid ClientId
+    {
+        get; set;
+    }
+    public string Reqion { get; set; } = null!;
+}
