+++ RateLimiter/Rules/IRuleCollection.cs
@@ -1,4 +1,5 @@
 using System.Collections;
+using RateLimiter.Models;
 
 namespace RateLimiter.Rules;
 public interface IRuleCollection
