--- RateLimiter/Token.cs
@@ -1,10 +1,0 @@
-using System;
-
-namespace RateLimiter;
-public class Token
-{
-    public Guid ClientId{
-        get; set;
-    }
-    public string Reqion{get;set;} = null!;
-}
